+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `soyhuce/next-ide-helper` will be documented in this file
 
 ## [Next release] - YYYY-MM-DD
 
+### Added
+
+- Laravel 9 support
+
 ## [0.6.3] - 2021-12-07
 
 ### Added